+++ quill-core/src/main/scala/io/getquill/quotation/Liftables.scala
@@ -105,23 +105,33 @@ trait Liftables {
     case AggregationOperator.`size` => q"$pack.AggregationOperator.`size`"
   }
 
+  implicit val renameableLiftable: Liftable[Renameable] = Liftable[Renameable] {
+    case Renameable.Fixed      => q"$pack.Renameable.Fixed"
+    case Renameable.ByStrategy => q"$pack.Renameable.ByStrategy"
+  }
+
+  implicit val visibilityLiftable: Liftable[Visibility] = Liftable[Visibility] {
+    case Visibility.Visible => q"$pack.Visibility.Visible"
+    case Visibility.Hidden  => q"$pack.Visibility.Hidden"
+  }
+
   implicit val queryLiftable: Liftable[Query] = Liftable[Query] {
-    case Entity(a, b)           => q"$pack.Entity($a, $b)"
-    case Filter(a, b, c)        => q"$pack.Filter($a, $b, $c)"
-    case Map(a, b, c)           => q"$pack.Map($a, $b, $c)"
-    case FlatMap(a, b, c)       => q"$pack.FlatMap($a, $b, $c)"
-    case ConcatMap(a, b, c)     => q"$pack.ConcatMap($a, $b, $c)"
-    case SortBy(a, b, c, d)     => q"$pack.SortBy($a, $b, $c, $d)"
-    case GroupBy(a, b, c)       => q"$pack.GroupBy($a, $b, $c)"
-    case Aggregation(a, b)      => q"$pack.Aggregation($a, $b)"
-    case Take(a, b)             => q"$pack.Take($a, $b)"
-    case Drop(a, b)             => q"$pack.Drop($a, $b)"
-    case Union(a, b)            => q"$pack.Union($a, $b)"
-    case UnionAll(a, b)         => q"$pack.UnionAll($a, $b)"
-    case Join(a, b, c, d, e, f) => q"$pack.Join($a, $b, $c, $d, $e, $f)"
-    case FlatJoin(a, b, c, d)   => q"$pack.FlatJoin($a, $b, $c, $d)"
-    case Distinct(a)            => q"$pack.Distinct($a)"
-    case Nested(a)              => q"$pack.Nested($a)"
+    case Entity.Opinionated(a, b, renameable) => q"$pack.Entity.Opinionated($a, $b, $renameable)"
+    case Filter(a, b, c)                      => q"$pack.Filter($a, $b, $c)"
+    case Map(a, b, c)                         => q"$pack.Map($a, $b, $c)"
+    case FlatMap(a, b, c)                     => q"$pack.FlatMap($a, $b, $c)"
+    case ConcatMap(a, b, c)                   => q"$pack.ConcatMap($a, $b, $c)"
+    case SortBy(a, b, c, d)                   => q"$pack.SortBy($a, $b, $c, $d)"
+    case GroupBy(a, b, c)                     => q"$pack.GroupBy($a, $b, $c)"
+    case Aggregation(a, b)                    => q"$pack.Aggregation($a, $b)"
+    case Take(a, b)                           => q"$pack.Take($a, $b)"
+    case Drop(a, b)                           => q"$pack.Drop($a, $b)"
+    case Union(a, b)                          => q"$pack.Union($a, $b)"
+    case UnionAll(a, b)                       => q"$pack.UnionAll($a, $b)"
+    case Join(a, b, c, d, e, f)               => q"$pack.Join($a, $b, $c, $d, $e, $f)"
+    case FlatJoin(a, b, c, d)                 => q"$pack.FlatJoin($a, $b, $c, $d)"
+    case Distinct(a)                          => q"$pack.Distinct($a)"
+    case Nested(a)                            => q"$pack.Nested($a)"
   }
 
   implicit val propertyAliasLiftable: Liftable[PropertyAlias] = Liftable[PropertyAlias] {
@@ -129,7 +139,7 @@ trait Liftables {
   }
 
   implicit val propertyLiftable: Liftable[Property] = Liftable[Property] {
-    case Property(a, b) => q"$pack.Property($a, $b)"
+    case Property.Opinionated(a, b, renameable, visibility) => q"$pack.Property.Opinionated($a, $b, $renameable, $visibility)"
   }
 
   implicit val orderingLiftable: Liftable[Ordering] = Liftable[Ordering] {