+++ quill-core/src/test/scala/io/getquill/ast/StatefulTransformerSpec.scala
@@ -1,6 +1,8 @@
 package io.getquill.ast
 
 import io.getquill.Spec
+import io.getquill.ast.Renameable.Fixed
+import io.getquill.ast.Visibility.Visible
 
 class StatefulTransformerSpec extends Spec {
 
@@ -305,6 +307,15 @@ class StatefulTransformerSpec extends Spec {
       Subject(Nil, Ident("a") -> Ident("a'"))(ast) match {
         case (at, att) =>
           at mustEqual Property(Ident("a'"), "b")
+          att.state mustEqual List(Ident("a"))
+      }
+    }
+
+    "property - fixed" in {
+      val ast: Ast = Property.Opinionated(Ident("a"), "b", Fixed, Visible)
+      Subject(Nil, Ident("a") -> Ident("a'"))(ast) match {
+        case (at, att) =>
+          at mustEqual Property.Opinionated(Ident("a'"), "b", Fixed, Visible)
           att.state mustEqual List(Ident("a"))
       }
     }