+++ quill-core/src/test/scala/io/getquill/ast/StatelessTransformerSpec.scala
@@ -1,6 +1,8 @@
 package io.getquill.ast
 
 import io.getquill.Spec
+import io.getquill.ast.Renameable.Fixed
+import io.getquill.ast.Visibility.Visible
 
 class StatelessTransformerSpec extends Spec {
 
@@ -165,6 +167,11 @@ class StatelessTransformerSpec extends Spec {
         Subject(Ident("a") -> Ident("a'"))(target) mustEqual
           OnConflict.Properties(List(Property(Ident("a'"), "b")))
       }
+      "properties - fixed" in {
+        val target: OnConflict.Target = OnConflict.Properties(List(Property.Opinionated(Ident("a"), "b", Fixed, Visible)))
+        Subject(Ident("a") -> Ident("a'"))(target) mustEqual
+          OnConflict.Properties(List(Property.Opinionated(Ident("a'"), "b", Fixed, Visible)))
+      }
     }
 
     "onConflict.action" - {
@@ -205,6 +212,12 @@ class StatelessTransformerSpec extends Spec {
       val ast: Ast = Property(Ident("a"), "b")
       Subject(Ident("a") -> Ident("a'"))(ast) mustEqual
         Property(Ident("a'"), "b")
+    }
+
+    "property - fixed" in {
+      val ast: Ast = Property.Opinionated(Ident("a"), "b", Fixed, Visible)
+      Subject(Ident("a") -> Ident("a'"))(ast) mustEqual
+        Property.Opinionated(Ident("a'"), "b", Fixed, Visible)
     }
 
     "infix" in {