+++ quill-orientdb/src/main/scala/io/getquill/context/orientdb/OrientDBIdiom.scala
@@ -29,7 +29,7 @@ trait OrientDBIdiom extends Idiom {
         case q: Query =>
           val sql = SqlQuery(q)
           VerifySqlQuery(sql).map(fail)
-          ExpandNestedQueries(sql, List()).token
+          new ExpandNestedQueries(naming)(sql, List()).token
         case other =>
           other.token
       }
@@ -234,7 +234,8 @@ trait OrientDBIdiom extends Idiom {
       case Property(ast, "isEmpty")   => stmt"${ast.token} IS NULL"
       case Property(ast, "nonEmpty")  => stmt"${ast.token} IS NOT NULL"
       case Property(ast, "isDefined") => stmt"${ast.token} IS NOT NULL"
-      case Property(ast, name)        => stmt"${strategy.column(name).token}"
+      case Property.Opinionated(ast, name, renameable, _) =>
+        renameable.fixedOr(name.token)(strategy.column(name).token)
     }
   }
 
@@ -268,10 +269,10 @@ trait OrientDBIdiom extends Idiom {
   implicit def actionTokenizer(implicit strategy: NamingStrategy): Tokenizer[Action] = {
 
     implicit def propertyTokenizer: Tokenizer[Property] = Tokenizer[Property] {
-      case Property(Property(_, name), "isEmpty")   => stmt"${strategy.column(name).token} IS NULL"
-      case Property(Property(_, name), "isDefined") => stmt"${strategy.column(name).token} IS NOT NULL"
-      case Property(Property(_, name), "nonEmpty")  => stmt"${strategy.column(name).token} IS NOT NULL"
-      case Property(_, name)                        => strategy.column(name).token
+      case Property(Property.Opinionated(_, name, renameable, _), "isEmpty")   => stmt"${renameable.fixedOr(name.token)(strategy.column(name).token)} IS NULL"
+      case Property(Property.Opinionated(_, name, renameable, _), "isDefined") => stmt"${renameable.fixedOr(name.token)(strategy.column(name).token)} IS NOT NULL"
+      case Property(Property.Opinionated(_, name, renameable, _), "nonEmpty")  => stmt"${renameable.fixedOr(name.token)(strategy.column(name).token)} IS NOT NULL"
+      case Property.Opinionated(_, name, renameable, _)                        => renameable.fixedOr(name.token)(strategy.column(name).token)
     }
 
     Tokenizer[Action] {
@@ -298,7 +299,8 @@ trait OrientDBIdiom extends Idiom {
   }
 
   implicit def entityTokenizer(implicit strategy: NamingStrategy): Tokenizer[Entity] = Tokenizer[Entity] {
-    case Entity(name, _) => strategy.table(name).token
+    case Entity.Opinionated(name, _, renameable) =>
+      renameable.fixedOr(name.token)(strategy.table(name).token)
   }
 
   protected def scopedTokenizer[A <: Ast](ast: A)(implicit token: Tokenizer[A]) =